requirements.txt
crewai==0.1.32
langgraph==0.0.15
langchain-community==0.0.14
python-dotenv==1.0.0
